add all deps

+++ src/static-analysis/cli.js
@@ -1,7 +1,8 @@
 const path = require('path');
 const { execSync } = require('child_process');
 const fs = require('fs-extra');
-const { getLocalRepoStats } = require('./getLocalRepoStats');
+const { getPatternflyStats } = require('./getPatternfyStats');
+const { getPackageStats, getAggregatePackageStats } = require('./getPackageStats');
 const repos = require('../../repos.json').repos;
 
 const statsDir = path.resolve(__dirname, '../../stats-static');
@@ -11,38 +12,58 @@ if (!fs.existsSync(statsDir)) {
   fs.mkdirSync(statsDir);
 }
 
+function collectPatternflyStats(argv) {
+  const date = new Date().toISOString();
+  const dir = `${statsDir}/${date.substr(0, 10)}`;
+  if (argv.c) {
+    fs.removeSync(tmpDir);
+  }
+  repos
+    .filter(repo => argv.p || !repo.private) // Only public repos unless flag passed
+    .forEach(repo => {
+      console.log(repo.name);
+      const repoName = repo.git.split('/').pop();
+      const tmpPath = `${tmpDir}/${repo.name}`;
+      if (!fs.existsSync(tmpPath)) {
+        const command = `git clone "${repo.git}" "${tmpPath}" --depth 1`;
+        console.log(command);
+        execSync(command);
+      }
+      const patternflyStats = getPatternflyStats(tmpPath);
+      const packageStats = getPackageStats(tmpPath, patternflyStats.name)
+      patternflyStats.repo = repo.git;
+      patternflyStats.name = repo.name || repoName;
+      patternflyStats.date = date;
+      if (argv.j) {
+        patternflyStats.dependencies = packageStats;
+      }
+
+      fs.outputFileSync(`${dir}/${repo.name}.json`, JSON.stringify(patternflyStats, null, 2));
+    });
+  
+  fs.outputFileSync(`${dir}/_all_dependencies.json`, JSON.stringify(getAggregatePackageStats(), null, 2));
+  console.log(`Collected stats for ${date} under ${dir}`);
+}
+
 require('yargs')
   .scriptName('repoStats')
   .usage('$0 ...flags')
   .command('collect', 'save stats locally', yargs => {
+    yargs.option('c', {
+      type: 'boolean',
+      default: 'false',
+      describe: 'whether to do a clean clone'
+    });
     yargs.option('p', {
       type: 'boolean',
       default: 'false',
       describe: 'whether to clone private repos'
-    })
-  }, argv => {
-    const date = new Date().toISOString();
-    const dir = `${statsDir}/${date.substr(0, 10)}`;
-    fs.removeSync(tmpDir);
-    repos
-      .filter(repo => argv.p || !repo.private) // Only public repos unless flag passed
-      .forEach(repo => {
-        const repoName = repo.git.split('/').pop();
-        const tmpPath = `${tmpDir}/${repo.name}`;
-        const command = `git clone "${repo.git}" "${tmpPath}" --depth 1`; 
-        console.log(command);
-        execSync(command);
-        const results = getLocalRepoStats(tmpPath);
-        results.repo = repo.git;
-        results.name = repo.name || repoName;
-        results.date = date;
-
-        if (!fs.existsSync(dir)) {
-          fs.mkdirSync(dir);
-        }
-        fs.writeFileSync(`${dir}/${repo.name}.json`, JSON.stringify(results, null, 2));
-      });
-    console.log(`Collected stats for ${date} under ${dir}`);
-  })
+    });
+    yargs.option('j', {
+      type: 'boolean',
+      default: 'false',
+      describe: 'whether to compile package.json stats'
+    });
+  }, collectPatternflyStats)
   .help()
-  .argv
+  .argv;
